rename test-data.js to run-status.js
